# Core-java_OpenAPI3.yaml
openapi: 3.0.1
info:
  title: vap-core
  description: 'Сервис базового функционала АС ''''Платформа видео-аналитики'''' '
  version: 0.1.0
servers:
  - url: /
...
paths:
...
  /nnModels:
    get:
      tags:
        - nnModels
      summary: Список нейросетевых моделей
      description: Возвращает список всех нейросетевых моделей из БД
      operationId: getNNMpdels
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/nnModel'
        404:
          description: NNModel not found
          content: {}
        405:
          description: Validation exception
          content: {}
    put:
      tags:
        - nnModels
      summary: Изменить нейросетевую модель
      description: Изменить метаданные новой нейросетевой модели в БД
      operationId: updateNNMpdel
      requestBody:
        description: Объект типа NNModel, который нужно добавить в БД
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/nnModel'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/nnModel'
            application/json:
              schema:
                $ref: '#/components/schemas/nnModel'
        400:
          description: Invalid id supplied
          content: {}
        404:
          description: NNModel not found
          content: {}
        405:
          description: Invalid nnModel
          content: {}
      x-codegen-request-body-name: body
    post:
      tags:
        - nnModels
      summary: Добавить нейросетевую модель
      description: Добавить метаданные новой нейросетевой модели в БД
      operationId: createNNModel
      requestBody:
        description: Объект типа NNModel, который нужно добавить в БД
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/nnModel'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/nnModel'
            application/json:
              schema:
                $ref: '#/components/schemas/nnModel'
        405:
          description: Invalid nnModel
          content: {}
      x-codegen-request-body-name: body
  /nnModels/{nnModelId}:
    delete:
      tags:
        - nnModels
      summary: Delete nnModel
      description: Deleting nnModel from DB.
      operationId: deleteNNModel
      parameters:
        - name: nnModelId
          in: path
          description: id нейросетевой модели, которую нужно удалить
          required: true
          schema:
            minimum: 1.0
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/nnModel'
            application/json:
              schema:
                $ref: '#/components/schemas/nnModel'
        400:
          description: Invalid id supplied
          content: {}
        404:
          description: nnModel not found
          content: {}
...
  /eventLogRecords:
    get:
      tags:
        - eventLogRecords
      summary: Список записей журнала происшествий
      description: Возвращает список всех  записей журнала происшествий из БД
      operationId: getEventLogRecords
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/eventLogRecord'
        404:
          description: BankUnits not found
          content: {}
    put:
      tags:
        - eventLogRecords
      summary: Сохранить запись журнала происшествий
      description: Сохранить изменения выбранной записи журнала происшествий в БД
      operationId: updateEventLogRecord
      requestBody:
        description: Объект типа EventLogRecord, который нужно добавить в БД
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/eventLogRecord'
          application/xml:
            schema:
              $ref: '#/components/schemas/eventLogRecord'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/eventLogRecord'
            application/json:
              schema:
                $ref: '#/components/schemas/eventLogRecord'
        400:
          description: Invalid id supplied
          content: {}
        404:
          description: EventLogRecord not found
          content: {}
        405:
          description: Invalid EventLogRecord
          content: {}
      x-codegen-request-body-name: body
    post:
      tags:
        - eventLogRecords
      summary: Добавить запись в журнал происшествий
      description: Добавить новую запись в таблицу журнал происшествий в БД
      operationId: placeEventLogRecord
      requestBody:
        description: Объект типа EventLogRecord, который нужно добавить в БД
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/eventLogRecord'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/eventLogRecord'
            application/json:
              schema:
                $ref: '#/components/schemas/eventLogRecord'
        405:
          description: Invalid EventLogRecord
          content: {}
      x-codegen-request-body-name: body
    delete:
      tags:
        - eventLogRecords
      summary: Удалить запись в журнале происшествий
      description: Удалить выбранную запись в журнале происшествий из БД
      operationId: deleteEventLogRecord
      parameters:
        - name: уventLogRecordId
          in: path
          description: id записи в журнале происшествий, которую нужно удалить
          required: true
          schema:
            minimum: 1.0
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/eventLogRecord'
            application/json:
              schema:
                $ref: '#/components/schemas/eventLogRecord'
        400:
          description: Invalid id supplied
          content: {}
        404:
          description: EventLogRecord not found
          content: {}
components:
  schemas:
...
    nnModel:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        fileName:
          type: string
        dockerImageName:
          type: string
        description:
          type: string
        comment:
          type: string
      xml:
        name: nnModel
...
    eventLogRecord:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cameraId:
          type: integer
          format: int64
        bankUnitId:
          type: integer
          format: int64
        nnModelId:
          type: integer
          format: int64
        eventTypeId:
          type: integer
          format: int64
        date:
          type: string
        timestamp:
          type: integer
          format: int64
        cadr:
          type: object
        comment:
          type: string
      xml:
        name: EventLogRecord
...